move 2930 tests
